Refactor unit tests as per the current changes
