+++ application/event_processor/tests/test_dos.py
@@ -2,18 +2,8 @@ from datetime import datetime, timezone
 from os import environ
 from random import choices
 from unittest.mock import patch
-
-from ..change_request import (
-    ADDRESS_CHANGE_KEY,
-    PHONE_CHANGE_KEY,
-    POSTCODE_CHANGE_KEY,
-    PUBLICNAME_CHANGE_KEY,
-    WEBSITE_CHANGE_KEY,
-)
 from ..dos import (
     DoSService,
-    add_address_to_change_request_if_not_equal,
-    add_field_to_change_request_if_not_equal,
     get_matching_dos_services,
     get_specified_opening_times_from_db,
 )
@@ -49,7 +39,7 @@ def test__init__publicname():
     dos_service.statusid = 1
     dos_service.publicname = "Test Public Name"
     expected_response = (
-        f"<uid={dos_service.uid} ods={dos_service.odscode} type={dos_service.typeid} "
+        f"<DoSService: uid={dos_service.uid} ods={dos_service.odscode} type={dos_service.typeid} "
         f"status={dos_service.statusid} name='{dos_service.publicname}'>"
     )
     # Assert
@@ -66,7 +56,7 @@ def test__init__name():
     dos_service.name = "Test Name"
     dos_service.publicname = None
     expected_response = (
-        f"<uid={dos_service.uid} ods={dos_service.odscode} type={dos_service.typeid} "
+        f"<DoSService: uid={dos_service.uid} ods={dos_service.odscode} type={dos_service.typeid} "
         f"status={dos_service.statusid} name='{dos_service.name}'>"
     )
     # Assert
@@ -83,136 +73,12 @@ def test__init__no_name():
     dos_service.publicname = None
     dos_service.name = None
     expected_response = (
-        f"<uid={dos_service.uid} ods={dos_service.odscode} type={dos_service.typeid} "
+        f"<DoSService: uid={dos_service.uid} ods={dos_service.odscode} type={dos_service.typeid} "
         f"status={dos_service.statusid} name='NO-VALID-NAME'>"
     )
     # Assert
     assert expected_response == str(dos_service), f"Should return {expected_response} string, actually: {dos_service}"
 
-
-def test_get_changes_same_data():
-    # Act
-    dos_service = dummy_dos_service()
-    nhs_kwargs = {
-        "Website": dos_service.web,
-        "Postcode": dos_service.postcode,
-        "Phone": dos_service.publicphone,
-        "OrganisationName": dos_service.publicname,
-        "Address1": dos_service.address,
-        "Address2": "",
-        "Address3": "",
-        "City": "",
-        "County": "",
-    }
-    nhs_entity = NHSEntity(nhs_kwargs)
-    # Act
-    response = dos_service.get_changes(nhs_entity)
-    # Assert
-    assert {} == response, f"Should return empty dict, actually: {response}"
-
-
-def test_get_changes_different_changes():
-    # Arrange
-    dos_service = dummy_dos_service()
-    website = "changed-website.com"
-    postcode = "TA1 TA1"
-    phone = "0123456789"
-    organisation_name = "changed-organisation-name"
-    address1 = "changed-address1"
-    address2 = "changed-address2"
-    address3 = "changed-address3"
-    city = "changed-city"
-    county = "changed-county"
-    nhs_kwargs = {
-        "Website": website,
-        "Postcode": postcode,
-        "Phone": phone,
-        "OrganisationName": organisation_name,
-        "Address1": address1,
-        "Address2": address2,
-        "Address3": address3,
-        "City": city,
-        "County": county,
-    }
-    nhs_entity = NHSEntity(nhs_kwargs)
-    expected_changes = {
-        ADDRESS_CHANGE_KEY: [address1, address2, address3, city, county],
-        PUBLICNAME_CHANGE_KEY: organisation_name,
-        WEBSITE_CHANGE_KEY: website,
-        POSTCODE_CHANGE_KEY: postcode,
-        PHONE_CHANGE_KEY: phone,
-    }
-    # Act
-    response = dos_service.get_changes(nhs_entity)
-    # Assert
-    assert expected_changes == response, f"Should return {expected_changes} dict, actually: {response}"
-
-
-def test_add_field_to_change_request_if_not_equal_is_equal():
-    # Arrange
-    changes = {}
-    nhs_uk_phone = "000000000"
-    dos_public_phone = "000000000"
-    # Act
-    actual_changes = add_field_to_change_request_if_not_equal(changes, "publicphone", dos_public_phone, nhs_uk_phone)
-    # Assert
-    assert changes == actual_changes, f"Should return {changes} dict, actually: {actual_changes}"
-
-
-def test_add_field_to_change_request_if_not_equal_not_equal():
-    # Arrange
-    changes = {}
-    nhs_uk_phone = "000000000"
-    dos_public_phone = "123456789"
-    expected_changes = {"publicphone": nhs_uk_phone}
-    # Act
-    actual_changes = add_field_to_change_request_if_not_equal(changes, "publicphone", dos_public_phone, nhs_uk_phone)
-    # Assert
-    assert expected_changes == actual_changes, f"Should return {expected_changes} dict, actually: {actual_changes}"
-
-
-def test_add_address_to_change_request_if_not_equal_is_equal():
-    # Arrange
-    changes = {}
-    nhs_uk_entity = NHSEntity({})
-    nhs_uk_entity.Address1 = "address1"
-    nhs_uk_entity.Address2 = "address2"
-    nhs_uk_entity.Address3 = "address3"
-    nhs_uk_entity.City = "city"
-    nhs_uk_entity.County = "county"
-    dos_address = (
-        f"{nhs_uk_entity.Address1}${nhs_uk_entity.Address2}$"
-        f"{nhs_uk_entity.Address3}${nhs_uk_entity.City}${nhs_uk_entity.County}"
-    )
-    # Act
-    actual_changes = add_address_to_change_request_if_not_equal(changes, "address", dos_address, nhs_uk_entity)
-    # Assert
-    assert changes == actual_changes, f"Should return {changes} dict, actually: {actual_changes}"
-
-
-def test_add_address_to_change_request_if_not_equal_not_equal():
-    # Arrange
-    changes = {}
-    nhs_uk_entity = NHSEntity({})
-    nhs_uk_entity.Address1 = "address1"
-    nhs_uk_entity.Address2 = "address2"
-    nhs_uk_entity.Address3 = "address3"
-    nhs_uk_entity.City = "city"
-    nhs_uk_entity.County = "county"
-    dos_address = "Test RD$Testown$Testshire"
-    expected_changes = {
-        ADDRESS_CHANGE_KEY: [
-            nhs_uk_entity.Address1,
-            nhs_uk_entity.Address2,
-            nhs_uk_entity.Address3,
-            nhs_uk_entity.City,
-            nhs_uk_entity.County,
-        ]
-    }
-    # Act
-    actual_changes = add_address_to_change_request_if_not_equal(changes, "address", dos_address, nhs_uk_entity)
-    # Assert
-    assert expected_changes == actual_changes, f"Should return {changes} dict, actually: {actual_changes}"
 
 
 @patch(f"{FILE_PATH}.connect")
@@ -253,7 +119,7 @@ def test_get_matching_dos_services_services_returned(mock_connect):
     ]
     mock_connect().cursor().fetchall.return_value = db_return
     odscode = "FQ038"
-    expected_response = "[<uid=159514725 ods=FQ038 type=13 status=1 name='My Pharmacy'>]"
+    expected_response = "[<DoSService: uid=159514725 ods=FQ038 type=13 status=1 name='My Pharmacy'>]"
     # Act
     response = get_matching_dos_services(odscode)
     # Assert
